refactor(devtools): split time_travel into 4 focused submodules
- Extract debugger logic to debugger.rs (605 lines)
- Extract types to types.rs (302 lines)
- Extract tests to tests.rs (637 lines)
- Keep module organization in mod.rs (25 lines)
- Original file was 1,580 lines
